requirements.txt
gigachat
gtts
pyaudio
pygame
vosk
python-dotenv